pyproject.toml
[tool.poetry]
name = "beauty-admin"
version = "0.1.0"
description = ""
authors = ["Serhii Lakodei <serhii.lakodei@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.27"
alembic = "^1.13.1"
gunicorn = "^21.2.0"
python-json-logger = "^2.0.7"
pydantic-settings = "^2.1.0"
geoalchemy2 = "^0.14.4"
flask-admin = "^1.6.1"
flask-sqlalchemy = "^3.1.1"
uvicorn = "^0.27.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
